Extract debounce book
